Simple Error checking
Looks at the 3 error bits of each reading to see if any of them are activated and only use's the readings that do not have any error bits.

My set up often returns the very long reading meaning that no good meserment where made for the sector as it is not spining at a controled speed with this code. But the error range is outside of the normal operating range and can there for be delt with as this range will only ever be the result of the lidar not reading properly.

I havent looked at the proccessing section enought to know if it is sensible or doing quite what i think but is is a start and my diff is hopefully easy to expand upon.