Quick Tweaks
Miss read the data sheet! only the last bite is a error bit but the second to last bit is also a signal strength warning so if ether of them are true then it ignores the data however i had had the third to last bit as a error bit as well but i turns out that on isnt a error or warning bit but a data bit so fixed to alow longer range

Oups :) more range now!